+++ packages/amplify-e2e-tests/src/__tests__/cdk/backends/custom-logic/reverse.ts
@@ -1,0 +1,13 @@
+/**
+ * Entry point for our lambda.
+ */
+export const handler = (event: any, _: any, callback: (errorMsg: string | null, responseMsg?: string | null) => void) => {
+  const messageToReverse: string = event.arguments.message;
+  if (!messageToReverse) {
+    callback("Didn't receive a `message` to reverse");
+  }
+  callback(
+    null,
+    messageToReverse.split('').reduce((acc, char) => char + acc, ''),
+  );
+};