+++ packages/amplify-e2e-tests/src/__tests__/schema-iterative-rollback-1.test.ts
@@ -18,7 +18,7 @@ import {
 } from 'amplify-category-api-e2e-core';
 
 // 30-45min
-describe('Iterative Rollback - add 2 @keys ', () => {
+describe('Iterative Rollback - add 2 @keys', () => {
   let projectDir: string;
   let appName: string;