+++ packages/amplify-graphql-api-construct/src/__tests__/__functional__/outputs.test.ts
@@ -23,9 +23,11 @@ describe('storeOutput', () => {
           "AWS::Amplify::Output": Object {
             "graphqlOutput": Object {
               "stackOutputs": Array [
+                "awsAppsyncApiId",
                 "awsAppsyncApiEndpoint",
                 "awsAppsyncAuthenticationType",
                 "awsAppsyncRegion",
+                "amplifyApiModelSchemaS3Uri",
                 "awsAppsyncApiKey",
               ],
               "version": "1",
@@ -69,6 +71,7 @@ describe('storeOutput', () => {
   describe('custom outputStorageStrategy', () => {
     const tokenRegex = /\$\{Token\[TOKEN\.\d+\]\}/;
     const awsRegionTokenRegex = /\$\{Token\[AWS\.Region\.\d+\]\}/;
+    const s3UriTokenRegex = /s3:\/\/\$\{Token\[TOKEN\.\d+\]\}\/.*/;
 
     const addBackendOutputEntry = jest.fn();
     const flush = jest.fn();
@@ -101,9 +104,11 @@ describe('storeOutput', () => {
         version: '1',
         payload: {
           awsAppsyncApiEndpoint: expect.stringMatching(tokenRegex),
+          awsAppsyncApiId: expect.stringMatching(tokenRegex),
           awsAppsyncApiKey: expect.stringMatching(tokenRegex),
           awsAppsyncAuthenticationType: 'API_KEY',
           awsAppsyncRegion: expect.stringMatching(awsRegionTokenRegex),
+          amplifyApiModelSchemaS3Uri: expect.stringMatching(s3UriTokenRegex),
         },
       });
       expect(flush).not.toBeCalled();
@@ -128,14 +133,15 @@ describe('storeOutput', () => {
         outputStorageStrategy,
       });
 
-      const template = Template.fromStack(stack);
       expect(addBackendOutputEntry).toBeCalledTimes(1);
       expect(addBackendOutputEntry).toBeCalledWith('graphqlOutput', {
         version: '1',
         payload: {
+          awsAppsyncApiId: expect.stringMatching(tokenRegex),
           awsAppsyncApiEndpoint: expect.stringMatching(tokenRegex),
           awsAppsyncAuthenticationType: 'OPENID_CONNECT',
           awsAppsyncRegion: expect.stringMatching(awsRegionTokenRegex),
+          amplifyApiModelSchemaS3Uri: expect.stringMatching(s3UriTokenRegex),
         },
       });
       expect(flush).not.toBeCalled();