+++ packages/graphql-transformers-e2e-tests/src/__tests__/MultiAuthV2TransformerWithFF.e2e.test.ts
@@ -425,7 +425,7 @@ afterAll(async () => {
   }
 });
 
-test("test 'public' authStrategy", async () => {
+test("'public' authStrategy", async () => {
   try {
     const createMutation = gql`
       mutation {
@@ -475,7 +475,7 @@ test("test 'public' authStrategy", async () => {
   }
 });
 
-test(`Test 'public' provider: 'iam' authStrategy`, async () => {
+test(`'public' provider: 'iam' authStrategy`, async () => {
   try {
     const createMutation = gql`
       mutation {
@@ -530,7 +530,7 @@ test(`Test 'public' provider: 'iam' authStrategy`, async () => {
   }
 });
 
-test(`Test 'private' authStrategy`, async () => {
+test(`'private' authStrategy`, async () => {
   try {
     const createMutation = gql`
       mutation {
@@ -585,7 +585,7 @@ test(`Test 'private' authStrategy`, async () => {
   }
 });
 
-test(`Test only allow private iam arn`, async () => {
+test(`only allow private iam arn`, async () => {
   try {
     const createMutation = gql`
       mutation {
@@ -664,7 +664,7 @@ test(`Test only allow private iam arn`, async () => {
   }
 });
 
-test(`Test 'private' provider: 'iam' authStrategy`, async () => {
+test(`'private' provider: 'iam' authStrategy`, async () => {
   // This test reuses the unauth role, but any IAM credentials would work
   // in real world scenarios, we've to see if provider override works.