+++ packages/graphql-transformers-e2e-tests/src/__tests__/VersionedModelTransformer.e2e.test.ts
@@ -114,7 +114,7 @@ afterAll(async () => {
 /**
  * Test queries below
  */
-test('Test createPost mutation', async () => {
+test('createPost mutation', async () => {
   const response = await GRAPHQL_CLIENT.query(
     `mutation {
         createPost(input: { title: "Hello, World!" }) {
@@ -134,7 +134,7 @@ test('Test createPost mutation', async () => {
   expect(response.data.createPost.version).toEqual(1);
 });
 
-test('Test updatePost mutation', async () => {
+test('updatePost mutation', async () => {
   const createResponse = await GRAPHQL_CLIENT.query(
     `mutation {
         createPost(input: { title: "Test Update" }) {
@@ -168,7 +168,7 @@ test('Test updatePost mutation', async () => {
   expect(updateResponse.data.updatePost.version).toEqual(2);
 });
 
-test('Test failed updatePost mutation with wrong version', async () => {
+test('failed updatePost mutation with wrong version', async () => {
   const createResponse = await GRAPHQL_CLIENT.query(
     `mutation {
         createPost(input: { title: "Test Update" }) {
@@ -203,7 +203,7 @@ test('Test failed updatePost mutation with wrong version', async () => {
   expect((updateResponse.errors[0] as any).errorType).toEqual('DynamoDB:ConditionalCheckFailedException');
 });
 
-test('Test deletePost mutation', async () => {
+test('deletePost mutation', async () => {
   const createResponse = await GRAPHQL_CLIENT.query(
     `mutation {
         createPost(input: { title: "Test Delete" }) {
@@ -233,7 +233,7 @@ test('Test deletePost mutation', async () => {
   expect(deleteResponse.data.deletePost.version).toEqual(createResponse.data.createPost.version);
 });
 
-test('Test deletePost mutation with wrong version', async () => {
+test('deletePost mutation with wrong version', async () => {
   const createResponse = await GRAPHQL_CLIENT.query(
     `mutation {
         createPost(input: { title: "Test Delete" }) {